refactored to separate buttonPanel which can be replaced in case its not attached
