+++ src/firmware/src/ui/ButtonPanel.h
@@ -1,21 +1,14 @@
-#ifndef __USERINTERFACE_H__
-#define __USERINTERFACE_H__
+#ifndef  __BUTTONPANEL_H__
+#define __BUTTONPANEL_H__
+
 #include <Wire.h>
 #include "Adafruit_MCP23017.h"
 
 #include <functional>
 #include <OneButton.h>
 
-#include "LedRing.h"
-#include "animations/WaveAnimation.h"
-#include "animations/GlowAnimation.h"
-
 #define BUTTON_PRESS_CALLBACK_TYPE std::function<void(ButtonEvent)>
 #define POWEROFF_STATE_CALLBACK_TYPE std::function<void()>
-
-struct RecordingProgressData {
-    int progress;
-};
 
 enum Action {
     Clicked,
@@ -36,7 +29,7 @@ struct ButtonContext {
     void* ui;
 };
 
-class UserInterface {
+class ButtonPanel {
 
     unsigned long lastInputScan;
     unsigned long inputScanInterval = 100;
@@ -46,28 +39,13 @@ class UserInterface {
     ButtonContext btnCtx2;
     ButtonContext btnCtx3;
 
-    LedRing ledRing;
-    WaveAnimation BlueProgressAnimation;
-    GlowAnimation WarmGlowAnimiation;
-    GlowAnimation SucessGlowAnimiation;
-    GlowAnimation WarningGlowAnimiation;
-    GlowAnimation ErrorGlowAnimiation;
-
     bool buttonStatus[3];
     bool isPowerOff = false;
-    float volume = 0.25;
 
     public:
 
-        UserInterface() : 
-            WarmGlowAnimiation(HUE_YELLOW, ANI_SPEED_FAST, 1),
-            SucessGlowAnimiation(HUE_GREEN, ANI_SPEED_FAST, 1),
-            WarningGlowAnimiation(HUE_ORANGE, ANI_SPEED_FAST, 1),
-            ErrorGlowAnimiation(HUE_RED, ANI_SPEED_FAST, 1)
-            { };
-
-        void setup();
-        void loop();
+        virtual void setup();
+        virtual void loop();
     
         BUTTON_PRESS_CALLBACK_TYPE buttonActionCallback;
         void onButtonEvent(BUTTON_PRESS_CALLBACK_TYPE callback) {
@@ -79,16 +57,8 @@ class UserInterface {
             this->powerOffCallback = callback;    
         }
 
-        float getVolume();
         virtual bool isPowerButtonOn();
-        void isBusy(bool);
-        void showRecordingProgress(int);
-        void showSuccess();
-        void showWarning();
-        void showError();
-        void showWelcome();
         void showHasNewMessageAt(int, bool);
-        void showAudioPlaying();
 };
 
-#endif // __USERINTERFACE_H__
+#endif //__BUTTONPANEL_H_LL
